move pagination functionality in separate component
