+++ src/components/CustomTable/Pagination.js
@@ -1,0 +1,37 @@
+import React from 'react';
+
+import IconButton from '@material-ui/core/IconButton';
+import { useTheme } from '@material-ui/core/styles';
+import KeyboardArrowLeft from '@material-ui/icons/KeyboardArrowLeft';
+import KeyboardArrowRight from '@material-ui/icons/KeyboardArrowRight';
+
+import { IconContainer } from './style';
+
+function TablePaginationActions(props) {
+  const theme = useTheme();
+  const { count, page, rowsPerPage, onPageChange } = props;
+
+  const handleBackButtonClick = (event) => {
+    onPageChange(event, page - 1);
+  };
+
+  const handleNextButtonClick = (event) => {
+    onPageChange(event, page + 1);
+  };
+
+  return (
+    <IconContainer>
+      <IconButton aria-label="previous page" disabled={page === 0} onClick={handleBackButtonClick}>
+        {theme.direction === 'rtl' ? <KeyboardArrowRight /> : <KeyboardArrowLeft />}
+      </IconButton>
+      <IconButton
+        aria-label="next page"
+        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
+        onClick={handleNextButtonClick}
+      >
+        {theme.direction === 'rtl' ? <KeyboardArrowLeft /> : <KeyboardArrowRight />}
+      </IconButton>
+    </IconContainer>
+  );
+}
+export default TablePaginationActions;
